faz os casos de teste para o ex6

+++ src/main/java/ListaExerciciosHerancaPolimorfismo/Ex6/ContaCorrente.java
@@ -2,19 +2,19 @@ package ListaExerciciosHerancaPolimorfismo.Ex6;
 
 public class ContaCorrente extends ContaBancaria {
 
-    public ContaCorrente(String numeroConta, String titular, double saldoInicial){
+    public ContaCorrente(String numeroConta, String titular, float saldoInicial){
         super(numeroConta,titular,saldoInicial);
     }
 
-    public double getTaxaManutencaoDeposito(){
-        return 0.50;
+    public float getTaxaManutencaoDeposito(){
+        return 0.50f;
     }
 
-    public double getTaxaManutencaoMensal(){
-        return 15.00;
+    public float getTaxaManutencaoMensal(){
+        return 15.00f;
     }
 
-    public void depositar(double valor){
+    public void depositar(float valor){
         if(valor <= 0){
             throw new IllegalArgumentException("Erro: Argumento valor deve ser positivo");
         }
